Add `NegatingUnsignedCast()` to correctly handle `std::numeric_limits<T>::min()` when converting a negated signed value to an unsigned type.
Cosmetics: do not unnecessarily rely on `size_t` being larger than `unsigned`.
PiperOrigin-RevId: 493600761

+++ riegeli/messages/message_parse.cc
@@ -188,7 +188,7 @@ absl::Status ParseFromString(absl::string_view src,
                              ParseOptions options) {
   bool parse_ok;
   if (ABSL_PREDICT_FALSE(src.size() >
-                         size_t{std::numeric_limits<int>::max()})) {
+                         unsigned{std::numeric_limits<int>::max()})) {
     parse_ok = false;
   } else if (!options.merge() && options.recursion_limit() ==
                                      google::protobuf::io::CodedInputStream::
